Develop

+++ packages/preload/src/services/installation-service.ts
@@ -1,17 +1,17 @@
 import * as fs from 'fs';
-import * as os from 'os';
 
 import path from 'path';
 import { loadConfiguration, saveConfiguration } from './configuration-service';
-import type { Mod } from '@interfaces/Mod';
-import type { AppConfiguration } from '@interfaces/AppConfiguration';
+import type { Mod } from '../../../../interfaces/Mod';
+import type { AppConfiguration } from '../../../../interfaces/AppConfiguration';
 
 import { buildPackage, extract, findStrings } from './pak-service';
-
+import { InstallationError } from '../../../../Errors/InstallationError';
 import {
   copyFiles,
   ensureDirectory,
   findFilesEndsWith,
+  getDocumentsPath,
   getFreeFileName,
   getLastExistingFileName,
   removeFileNameExtension,
@@ -19,17 +19,20 @@ import {
 } from './file-service';
 import { loadMods } from './mod-service';
 import { updateProgressBar } from './progress-service';
-
-const G3_DOCUMENTS_PATH = path.join(os.homedir(), 'Documents', 'gothic3');
+import { UTF8 } from '../../../../utils/constants';
+import { loggerError, loggerInfo } from './logger-service';
+import { getMessage } from '../../../../utils/messages';
+import { showAlert } from './alert-service';
+import { ConfigurationError } from '../../../../Errors/ConfigurationError';
 
 const APP_PATH = path.resolve();
 const STATIC_FILES_PATH = path.join(APP_PATH, 'Static');
 const PRESET_FILES_PATH = path.join(APP_PATH, 'Presets');
-const STATIC_FILE_MOD_EXTENSTION = '0x';
+const STATIC_FILE_MOD_EXTENSION = '0x';
 const STRINGTABLE_FILENAME = 'stringtable.ini';
 // const STRINGTABLEMOD_FILENAME = 'stringtablemod.ini';
 
-const MOD_EXTENSTIONS = ['mod', 'nod'];
+const MOD_EXTENSIONS = ['mod', 'nod'];
 const DLL_EXTENSION = 'dll';
 const INI_EXTENSION = 'ini';
 const SAVE_EXTENSION = 'g3savcpx';
@@ -39,12 +42,12 @@ const SHADER = 'Shader.Cache';
 const WRLDATASC = 'G3_World_01.wrldatasc';
 
 const STRINGTABLE_ENCODING = 'utf16le';
-const WRLDATASC_ENCODING = 'utf-8';
 
 export async function installMods(modIds: string[], preset?: string): Promise<string> {
   return new Promise((resolve, reject) => {
     (async () => {
-      const configuration: AppConfiguration = (await loadConfiguration()) as AppConfiguration;
+      const configuration = await loadConfiguration();
+      if (!configuration) throw new InstallationError(getMessage('CONFIG_NOT_FOUND'));
       const filesDictionary = prepareFilesDictionary();
       const scriptFiles: string[] = [];
       const iniFiles: string[] = [];
@@ -56,19 +59,27 @@ export async function installMods(modIds: string[], preset?: string): Promise<st
       const mods = allMods.filter(mod => modIds.includes(mod.id));
 
       try {
+        if (!modIds.length) throw new InstallationError(getMessage('NO_MODS_SELECTED'));
         const startTime = performance.now();
+        loggerInfo(getMessage('INSTALLATION_START', { num: modIds.length.toString() }));
+        loggerInfo(getMessage('INSTALLATION_MOD_LIST', { mods: modIds.join(', ') }));
         await deleteMods();
         await moveSaves();
 
-        await moveSplash(configuration, preset);
-        await moveShader(preset);
+        if (preset) {
+          await moveSplash(configuration, preset);
+          await moveShader(preset);
+        }
+
+        if (!fs.existsSync(dataPath))
+          throw new InstallationError(getMessage('DATA_DIR_DOESNT_EXIST'));
 
         for (let i = 0; i < mods.length; i++) {
           updateProgressBar('progress.searchMods', i, mods.length);
-          for (const extension of MOD_EXTENSTIONS) {
+          for (const extension of MOD_EXTENSIONS) {
             const files = findFilesEndsWith(
               mods[i].path,
-              `${extension[0]}${STATIC_FILE_MOD_EXTENSTION}`,
+              `${extension[0]}${STATIC_FILE_MOD_EXTENSION}`,
             );
             if (!filesDictionary[extension]) {
               filesDictionary[extension] = [];
@@ -80,36 +91,45 @@ export async function installMods(modIds: string[], preset?: string): Promise<st
           const inis = findFilesEndsWith(mods[i].path, INI_EXTENSION);
           iniFiles.push(...inis);
         }
-
         appendFakeFiles(filesDictionary);
-
         for (const key in filesDictionary) {
-          {
-            await copyFiles(dataPath, key, filesDictionary[key], createdFiles);
-          }
+          await copyFiles(dataPath, key, filesDictionary[key], createdFiles);
         }
+
         await copyScriptsFiles(scriptsPath, scriptFiles, createdFiles);
         await copyScriptsFiles(iniPath, iniFiles, createdFiles);
-
         await buildStringTable(dataPath, mods, createdFiles);
         await buildWrldatasc(dataPath, mods, createdFiles);
         const endTime = performance.now();
         const time = (endTime - startTime) / 1000;
 
+        updateProgressBar('progress.saveConfiguration', 0, 2);
+
         configuration.installedMods = mods.map(mod => mod.id);
-        //Get rid of possible duplicates
         configuration.preset = preset ? preset : undefined;
+        //Get rid of possible duplicates
         configuration.filesCreated = Array.from(new Set(createdFiles));
 
         await saveConfiguration(configuration);
-
-        resolve(time.toFixed(3));
+        loggerInfo(getMessage('INSTALLATION_COMPLETE', { num: modIds.length.toString() }));
+        resolve(time.toFixed(2));
       } catch (error) {
-        // Remove copied files
-        alert(error);
-        for (let i = 0; i < createdFiles.length; i++) {
-          updateProgressBar('progress.delete', i, createdFiles.length);
-          fs.unlinkSync(createdFiles[i]);
+        if (error instanceof Error) {
+          showAlert('modal.error', getMessage('CHECK_LOG_FILE'), 'error');
+          loggerError(getMessage('INSTALLATION_FAIL'));
+          loggerInfo(getMessage('REVERT_INSTALLATION_CHANGES'));
+          // Remove copied files
+          if (createdFiles.length) {
+            for (let i = 0; i < createdFiles.length; i++) {
+              loggerInfo(getMessage('FILE_DELETING', { path: createdFiles[i] }));
+              updateProgressBar('progress.delete', i, createdFiles.length);
+              fs.unlinkSync(createdFiles[i]);
+              loggerInfo(getMessage('FILE_DELETED', { path: createdFiles[i] }));
+            }
+            loggerInfo(getMessage('REVERT_COMPLETE'));
+          } else {
+            loggerInfo(getMessage('REVERT_NOTHING_TO_DO'));
+          }
         }
         reject(error);
       }
@@ -120,7 +140,7 @@ export async function installMods(modIds: string[], preset?: string): Promise<st
 function prepareFilesDictionary(): Record<string, string[]> {
   const filesDictionary: Record<string, string[]> = {};
 
-  for (const extension of MOD_EXTENSTIONS) {
+  for (const extension of MOD_EXTENSIONS) {
     filesDictionary[extension] = [];
   }
   return filesDictionary;
@@ -143,18 +163,25 @@ async function copyScriptsFiles(
       continue;
     }
     const newFilePath = path.join(destinationPath, fileName);
+
+    loggerInfo(getMessage('COPY_FILE_FROM_TO', { src: filePath, dst: newFilePath }));
     await fs.promises.copyFile(filePath, newFilePath);
+    loggerInfo(getMessage('COPY_FILE_FROM_TO_COMPLETE', { src: filePath, dst: newFilePath }));
     createdFiles.push(newFilePath);
   }
 }
 
 export async function deleteMods(): Promise<void> {
-  const configuration: AppConfiguration = (await loadConfiguration()) as AppConfiguration;
-
+  const configuration = await loadConfiguration();
+  if (!configuration) throw new ConfigurationError(getMessage('MISSING_CONFIGURATION'));
   const filesCount = configuration.filesCreated.length;
+
+  if (!filesCount) return;
   for (let i = 0; i < filesCount; i++) {
     updateProgressBar('progress.delete', i, filesCount);
+    loggerInfo(getMessage('FILE_DELETING', { path: configuration.filesCreated[i] }));
     fs.unlinkSync(configuration.filesCreated[i]);
+    loggerInfo(getMessage('FILE_DELETED', { path: configuration.filesCreated[i] }));
   }
 
   configuration.installedMods = [];
@@ -163,7 +190,7 @@ export async function deleteMods(): Promise<void> {
   await saveConfiguration(configuration);
 }
 
-function appendFakeFiles(dictionary: Record<string, string[]>): void {
+export function appendFakeFiles(dictionary: Record<string, string[]>): void {
   dictionary['mod'].push(path.join(STATIC_FILES_PATH, 'Projects_compiled.m0x'));
   dictionary['nod'].push(path.join(STATIC_FILES_PATH, 'Projects_compiled.n0x'));
 }
@@ -190,13 +217,13 @@ async function buildStringTable(
   fs.rmSync(tempDir, { recursive: true, force: true });
 }
 
-async function mergeStringTables(
+export async function mergeStringTables(
   gothicDataPath: string,
   mods: Mod[],
   originalStringTable: string,
 ): Promise<void> {
   const tempDir = path.join(gothicDataPath, 'temp');
-  await ensureDirectory(tempDir);
+  ensureDirectory(tempDir);
   await extract(originalStringTable, [STRINGTABLE_FILENAME], tempDir);
   const stringTable = path.join(tempDir, STRINGTABLE_FILENAME);
   const locAdminRevision = '[LocAdmin_Revisions]';
@@ -209,8 +236,9 @@ async function mergeStringTables(
     flag: 'w',
   });
 
-  mods.forEach(element => {
-    const modStringTable = path.join(element.path, STRINGTABLE_FILENAME);
+  for (let i = 0; i < mods.length; i++) {
+    updateProgressBar('progress.buildStringtable', i, mods.length);
+    const modStringTable = path.join(mods[i].path, STRINGTABLE_FILENAME);
     if (fs.existsSync(modStringTable)) {
       const modStringTableContent = fs
         .readFileSync(modStringTable, {
@@ -223,7 +251,7 @@ async function mergeStringTables(
         flag: 'a',
       });
     }
-  });
+  }
 
   fs.appendFileSync(stringTable, `\n${locAdminRevision}`, {
     encoding: STRINGTABLE_ENCODING,
@@ -231,9 +259,10 @@ async function mergeStringTables(
   });
 }
 
-async function buildWrldatasc(gothicDataPath: string, mods: Mod[], createdFiles: string[]) {
+export async function buildWrldatasc(gothicDataPath: string, mods: Mod[], createdFiles: string[]) {
   const wrldataPath = path.join(STATIC_FILES_PATH, WRLDATASC);
-  if (!fs.existsSync(wrldataPath)) throw new Error('No wrldatasc file');
+  if (!mods.length) throw new InstallationError('No mods selected');
+  if (!fs.existsSync(wrldataPath)) throw new InstallationError('No wrldatasc file');
   const outputFileName = await getFreeFileName(gothicDataPath, 'projects_compiled', 'mod');
   const outputFilePath = path.join(gothicDataPath, outputFileName);
   const lastExistingFileName = await getLastExistingFileName(
@@ -242,51 +271,62 @@ async function buildWrldatasc(gothicDataPath: string, mods: Mod[], createdFiles:
     'mod',
   );
   const lastExistingFileNamePath = path.join(gothicDataPath, lastExistingFileName);
-  if (fs.existsSync(outputFilePath)) fs.unlinkSync(outputFilePath);
   const tempDir = path.join(gothicDataPath, 'temp');
-  await ensureDirectory(tempDir);
-  await ensureDirectory(path.join(tempDir, 'G3_World_01'));
+  ensureDirectory(tempDir);
+  ensureDirectory(path.join(tempDir, 'G3_World_01'));
   const tempWrldataPath = path.join(tempDir, 'G3_World_01', WRLDATASC);
   fs.copyFileSync(wrldataPath, tempWrldataPath);
-  mods.forEach(mod => {
-    const wrldataMod = path.join(mod.path, WRLDATASC);
+  for (let i = 0; i < mods.length; i++) {
+    updateProgressBar('progress.buildStringtable', i, mods.length);
+    const wrldataMod = path.join(mods[i].path, WRLDATASC);
     if (fs.existsSync(wrldataMod)) {
       const wrldataModContent = fs.readFileSync(wrldataMod, {
-        encoding: WRLDATASC_ENCODING,
+        encoding: UTF8,
         flag: 'r',
       });
       const sectors = wrldataModContent.replace(/^[\s\S]*?\[Sector\.List\]/, '');
       fs.appendFileSync(tempWrldataPath, sectors, {
-        encoding: WRLDATASC_ENCODING,
+        encoding: UTF8,
         flag: 'a',
       });
     }
-  });
+  }
 
   await buildPackage(tempDir, outputFilePath);
   fs.rmSync(tempDir, { recursive: true, force: true });
-  if (path.extname(outputFilePath) == 'mod') return;
   swapFileNames(outputFilePath, lastExistingFileNamePath);
+
   createdFiles.push(outputFilePath);
 }
 
-async function moveSaves() {
-  const newModsFolder = getNextSaveDirectoryName();
+export async function moveSaves() {
+  const newModsFolder = await getNextSaveDirectoryName();
+  const G3_DOCUMENTS_PATH = path.join(await getDocumentsPath(), 'gothic3');
+
   const newModsFolderPath = path.join(G3_DOCUMENTS_PATH, newModsFolder);
-  const oldFilesPaths = getOldModsFiles();
+  const oldFilesPaths = await getOldModsFiles();
+  loggerInfo(getMessage('MOVE_SAVES_START'));
   if (oldFilesPaths.length === 0) {
+    loggerInfo(getMessage('MOVE_SAVES_NOTHING_TO_MOVE'));
     return;
   }
 
-  await ensureDirectory(newModsFolderPath);
-  const newFilesPaths = getNewModsFilesPaths(oldFilesPaths, newModsFolder);
+  ensureDirectory(newModsFolderPath);
+  const newFilesPaths = await getNewModsFilesPaths(oldFilesPaths, newModsFolder);
   for (let i = 0; i < oldFilesPaths.length; i++) {
+    loggerInfo(getMessage('MOVE_SAVES_FROM_TO', { src: oldFilesPaths[i], dst: newFilesPaths[i] }));
     updateProgressBar('progress.moveOldSaves', i, oldFilesPaths.length);
     await fs.promises.rename(oldFilesPaths[i], newFilesPaths[i]);
+    loggerInfo(
+      getMessage('MOVE_SAVES_FROM_TO_COMPLETE', { src: oldFilesPaths[i], dst: newFilesPaths[i] }),
+    );
   }
+  loggerInfo(getMessage('MOVE_SAVES_COMPLETE'));
 }
 
-function getOldModsFiles() {
+export async function getOldModsFiles() {
+  const G3_DOCUMENTS_PATH = path.join(await getDocumentsPath(), 'gothic3');
+
   const saves = findFilesEndsWith(G3_DOCUMENTS_PATH, SAVE_EXTENSION);
   const savesDat = findFilesEndsWith(G3_DOCUMENTS_PATH, SAVEDAT_EXTENSION);
   const shader = findFilesEndsWith(G3_DOCUMENTS_PATH, 'Cache');
@@ -294,8 +334,10 @@ function getOldModsFiles() {
   return saves.concat(savesDat).concat(shader);
 }
 
-function getNextSaveDirectoryName(): string {
+export async function getNextSaveDirectoryName(): Promise<string> {
   let maxNumber = -1;
+  const G3_DOCUMENTS_PATH = path.join(await getDocumentsPath(), 'gothic3');
+
   const directories = fs.readdirSync(G3_DOCUMENTS_PATH);
 
   for (const directory of directories) {
@@ -315,8 +357,10 @@ function getNextSaveDirectoryName(): string {
   return `Mods${nextNumber.toString()}`;
 }
 
-function getNewModsFilesPaths(files: string[], destDirectory: string) {
+export async function getNewModsFilesPaths(files: string[], destDirectory: string) {
   const newFilesPaths: string[] = [];
+  const documents = await getDocumentsPath();
+  const G3_DOCUMENTS_PATH = path.join(documents, 'gothic3');
   for (const file of files) {
     const fileName = path.basename(file);
     const newFilePath = path.join(G3_DOCUMENTS_PATH, destDirectory, fileName);
@@ -325,28 +369,32 @@ function getNewModsFilesPaths(files: string[], destDirectory: string) {
   return newFilesPaths;
 }
 
-async function moveSplash(configuration: AppConfiguration, presetName?: string) {
+export async function moveSplash(configuration: AppConfiguration, presetName: string) {
   let splash = path.join(STATIC_FILES_PATH, SPLASH);
   if (presetName && fs.existsSync(path.join(PRESET_FILES_PATH, presetName, SPLASH))) {
     splash = path.join(PRESET_FILES_PATH, presetName, SPLASH);
   }
 
   if (!fs.existsSync(splash)) {
+    loggerInfo(getMessage('COPY_SPLASH_NOT_FOUND', { preset: presetName }));
     return;
   }
-
+  loggerInfo(getMessage('COPY_SPLASH_START'));
   const splashDest = path.join(configuration.gothicPath, SPLASH);
   fs.copyFileSync(splash, splashDest);
+  loggerInfo(getMessage('COPY_SPLASH_COMPLETE'));
 }
 
-async function moveShader(presetName?: string) {
-  if (!presetName) {
-    return;
-  }
+export async function moveShader(presetName: string) {
+  const G3_DOCUMENTS_PATH = path.join(await getDocumentsPath(), 'gothic3');
   const shader = path.join(PRESET_FILES_PATH, presetName, SHADER);
+
   if (!fs.existsSync(shader)) {
+    loggerInfo(getMessage('COPY_SHADER_NOT_FOUND', { preset: presetName }));
     return;
   }
   const shaderDest = path.join(G3_DOCUMENTS_PATH, SHADER);
+  loggerInfo(getMessage('COPY_SHADER_START'));
   fs.copyFileSync(shader, shaderDest);
+  loggerInfo(getMessage('COPY_SHADER_COMPLETE'));
 }
